pyproject.toml
[project]
name = "imgenx"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.4",
    "python-dotenv>=1.1.1",
    "volcengine-python-sdk[ark]>=4.0.22",
    "requests>=2.25.0",
    "pillow>=12.0.0",
]

[project.urls]
repository = "https://github.com/NewToolAI/imgenx"

[project.scripts]
imgenx = "imgenx.main:run"
